Add Services to database
The Services table holds key information for Pedalspace's functionality. Each entry in the Services table will describe a specific service offered by a specific mechanic.

Service listings contain the name of the service, the price, and a reference to the ID of the mechanic they belong to (mechanics have many services). This means that there may be several distinct entries for services with the same name, because they will be offered by different mechanics (who may for example charge different prices for the same service).