Add Reviews to database
Reviews are the way that Pedalspace community members will build trust between each other. After every completed job, the cyclist and the mechanic will each leave a review for the other which will show up on their profile.

This means that there are two subclasses of reviews to be stored in the Reviews table (single table inheritance). Both types will belong to both a cyclist and a mechanic (one party leaves the review, the other receives it).

Reviews of cyclists will contain a field to indicate whether or not they paid the agreed-upon price, and may also contain comments (optional text).

Reviews of mechanics will contain fields to indicate whether or not they provided the expected service at the expected price, and may also contain comments (optional text).

Both types of reviews also contain a numeric rating meant to give an overview of interaction as a whole.