Added controllers for Users, Bookings and Reviews
The intention is for these 3 controllers to be a parent of the child Namespaced controllers Users::Cyclists/Users::Mechanics along with Cyclists::Bookings / Mechanic::Bookings as well as Cyclists::Reviews / Mechanics::Reviews. We are unsure at this point if these controllers will be utilized, but we are generating them in the case where use is necessary.

Along with the Controllers for Users/Bookings/Reviews, stylesheets and test files have also been generated