Add Requested Services to database
Requested services will link each booking that a cyclist makes with a mechanic, to the services requested in that booking. Because both cyclists and mechanics will have access to all of their past and current bookings, this  will allow both parties to keep a record of which services were booked/performed and when. This table acts as a join table, allowing services to have many bookings (through requested services) and bookings to have many services (through requested services).