Add scoped comments resources for mechanics bookings
Comments provide a way for mechanics to update bookings, by explaining services completed, or added or cancelled from the job.

These routes are created as a nested resource, because comments only make sense in the context of a specific booking, and only mechanics should be able to interact with them. Comments should not be able to be deleted, because they provide historical context and reminders for both parties about what work has actually been completed and why.

While this nesting creates some long/complex routes, we have decided that this is currently the best option for keeping resources and functionality logically organized.