Add actions for updating Reviews of Cyclists
The 'edit' action simply returns the review instance to be edited, which is found by taking the id from the params.

The 'update' action updates the review by taking in the same strong params as are used in creating a new review. If the update is successful, redirect to the show the cyclist's profile, which is where the reviews should be displayed.