Add actions for creating mechanics Services
The 'new' action simply returns a new (empty) instantiation of the Service model.

The 'create' action takes in parameters from the session and the user to create a new service:
- The ID of the mechanic adding a service is taken from the session
- Strong params are defined to accept params containing details for the service (service name & price)
- The service is built via the mechanic object which automatically attaches it to that mechanic

If the service is successfully saved to the database, redirect to the mechanic's list of services.
If it isn't successfully saved, re-render the form for creating a new service (ideally the initial details entered should be preserved and error messages should be displayed).