Add actions for updating mechanics services
The 'edit' action simply returns the service instance to be edited, which is found by taking the id from the params.

The 'update' action updates the service by taking in the same strong params as are used in creating a new service. If the update is successful, redirect to the show the mechanic's list of services; otherwise, re-render the edit view.