Pluralize routes for reviews
This commit pluralizes the scoped routes created for reviews of cyclists and reviews of mechanics, so that these routes will talk to the existing controller set-up.

Note that it may have made more sense to to use the singular form (Review of Cyclist, or Reviews of Cyclist), but given the current timeline for the project, it currently makes more sense to make the smallest change possible that will make things functional and allow us to move forward.