Add new view for Bookings
The new view allows a Cyclist to submit a new Booking request to a Mechanic.  It sets the Cyclist through sesson[:id] and Mechanic through :mechanic_id in the route.  The description is submitted through a free text field, and the default status is set automatically through a hidden field.  The Cyclist indicates which services they want via checkboxes, which is transmitted via an array of service ids.