Fix scoped routes for reviews
Because actions concerning reviews only make sense in the context of the cyclist/mechanic that they are *about*, other routes are superfluous.

The only actions that should be taken wrt reviews are creating them and updating them, both of which make the most sense when done in the context of the person they are about rather than the person who wrote them (though obviously only the person who wrote them should have access to those actions - this means a filter will be required).
Reviews will only be viewed on the profile of the person they are attached to, which does not require a specific action.