Added a yield for header content Added png image to nav. Added login form.
The header content was removed from index.html.erb and given a <%yield> for application.html.erb
What his means, is that the application.html.erb will yield for .ERB header content in each view.
The embedded header content is denoted by:
<% content_for :header do -%>
<% end %>

As a result, a <body> and <head> tag were removed from index.html.erb

The navbar content was modified for a new view rendering.
A png file was added to the navbar on index.html.erb, but is not yet renderd.

In order to render  jQuery, the turbolinks gem needed to be added in Gemfile.