Add initial ElasticSearch to Mechanic model
This commit adds the first basic layer of ElasticSearch functionality to the Mechanic model. The model now requires and includes the ElasticSearch Model module, which adds all the default ElasticSearch methods to the model (including, for example, 'search').
The line to include Elasticsearch::Model::Callbacks sets callbacks that update Elastic Search whenever the model is changed.
The line at calling Mechanic.import at the bottom of the file sends all the Mechanic information to ElasticSearch every time the app is started.