Add basic Elastic settings to Mechanic model
This commit adds basic ElasticSearch indexing settings. Most of the code added is example code from the docs - the only change was to change the column being indexed to :first_name from the default (:title). This code tells Elastic to index the first_name column and use the 'english' analyzer.