Add initial location search to mechanics
Add new clause to search function to restrict search results to a city. The new clause relies on a new proc that searches mechanics whose address puts them in the city listed in the params (if given) and returns only distinct results.

Note - this is an interim measure: location-based searching should rely on a maps integration and ideally should default to finding the user's location from the device they are using instead of relying on them to type it in.