Update Mechanic controller to take search_params
This commit sets conditional check in the mechanics#index action to verify whether acceptable search parameters can be found in the params hash: :city, :service, or :price. If any of those are present, it passes the search_params to the Mechanics' search method and returns the results. The protected method that defines which params are permitted as search params is now updated to be consistent with what the model's search method is expecting.
If none of those params are present, the entire list of mechanics is returned.