Add Devise registration controller
Devise uses a custom controller for the creation for users.  This controller overrides default behavior, allowing for Cyclist and Mechanic users to be created from a single page and correctly routed to their respective profile pages.

Users are not currently logged in automatically after sign in.  This feature will be added later.