Refactor mechanic services search: split query into terms
This commit refactors the method that handles searching services to improve the handling of multi-word searches. Using a protected helper method, the service search query is split on spaces and each term is converted into a PG-friendly regular expression. The search can then check for the words in the query in any order, in any place in the service_name field. Note that this version of the search uses 'LIKE ALL' instead of 'LIKE ANY' so that every word in the search query must be matched, not just one: this is done by design, so that searches for 'build wheel' won't return mechanics who only offer 'true wheel'.