Created Mechanic Profile view page. Updated Cyclist Profile View page.
The mechanic profile view page was rendered from the previous version of cyclist view.
Embedded logic was added to some componenets of the view, while other logic was  left as text - using  <(space)%= notation.

The cyclist/show view was altered to match the logic from the Wireframes.
Some ERB logic was rendered in cyclist view, but again same as above -some was left as text.

The cyclist controller was touched, just to comment out the controller for initial viewing purposes.