Add location-aware map to homepage
This commit replaces the placeholder map on the home page with a location-aware map. It uses example code provided as-is from the Google Maps API reference docs.
The only custom element is the zoom level, which is set to 12. This is a closer zoom than the default. It was chosen because it displays a more reasonable area to consider when looking for a nearby mechanic, since the user will likely be seeking that mechanic on a bike and may not want to travel a long distance.