Add location columns to Users table in db
In order to place mechanics on the map, their address needs to be 'geocoded' - ie, converted to latitude-longitude coordinates. Since geocoding is a resource-intensive and API-intensive task, it makes the most sense by far to store that information in the database once it has bee calculated. Therefore, this migration adds two columns to the Users table to store the latitude and longitude calculated for each mechanic user.