Add geocoding to Mechanic model
In order to display mechanics on a map, their addresses need to be converted to geocoded latitude/longitude coordinates.

This commit adds the following:
- A helper method to get the mechanic's full address from their separate address fields
- A 'geocoded_by' call that tells the Geocoder gem what information to look at to find the coordinates
- An 'after_validation' step that tells the Geocoder gem to automatically geocode each Mechanic after it has been validated.
- A commented-out note in the search-by-city method with a reminder of the syntax to be added when the search is modified to take the location into account more intelligently than checking the 'city' string

NOTE that geocoding is a time- and resource-intensive task and *should not* be done after every save. As soon as possible, an 'if' check should be added to the 'after_validation :geocode' step so that it only performs the geocoding if the address information is present and has changed.