Replace dummy marker with Mechanic markers
This commit replaces the dummy (test) marker function with a function to display markers for all mechanics returned from the database.

The new code makes an AJAX call to the database to retrieve the list of mechanics as JSON. In the 'success' function for the AJAX call, it loops through the results and sets a new maps marker for each one. It relies on each mechanic object to have both a latitude and longitude attribute.

First note: There is currently no error handling in case of bad data.

Second note: This function was initially written using the $.getJSON method, but for unexplained reasons, we were unable to assign the results from that call to a variable in order to work with them.