Add location-based search scope to Mechanics
This commit adds a new search scope to the Mechanic's #search method. If data is passed in for the user's location from params, the search will now start by finding results near that location. This method relies on the 'near' method provided by the Geocoder gem.
The strong params defined for search_params now also permits 'location' to be passed in.