Add search functionality to Pages controller
Because the search results will be displayed at /pages/home, the Pages controller needs to be able to handle search requests. For simplicity's sake, this commit adds search handling directly to the Pages#home action. The search handling code is copied directly from the Mechanics controller where it was originally defined, with two small changes:
- Add a check for 'location' in the params
- Permit 'location' in the strong params