Convert search field to form on home page
This commit removes the previous text input field and replaces it with an erb form_tag as a search box.
- For now, text entered in the search field will be treated as a search for a service
- The user's location is passed in via a hidden field to restrict search results to 'near them'

Note - efforts were made to preserve existing styling, but the icon now appears outside the text field.
Also note - the map still focuses on Victoria even if the search query should be focused elsewhere (tb updated).